requirements.txt
azure-cognitiveservices-speech>=1.38.0
requests>=2.31.0